+++ src/components/MovieForm.tsx
@@ -1,0 +1,137 @@
+import React from "react";
+import { type Theme } from "../styles/theme";
+
+interface MovieFormProps {
+  newTitle: string;
+  newDirector: string;
+  newYear: number | "";
+  newGenre: string;
+  newRating: number | "";
+  setNewTitle: (value: string) => void;
+  setNewDirector: (value: string) => void;
+  setNewYear: (value: number | "") => void;
+  setNewGenre: (value: string) => void;
+  setNewRating: (value: number | "") => void;
+  onSubmit: (e: React.FormEvent) => void;
+  currentTheme: Theme;
+}
+
+const MovieForm: React.FC<MovieFormProps> = ({
+  newTitle,
+  newDirector,
+  newYear,
+  newGenre,
+  newRating,
+  setNewTitle,
+  setNewDirector,
+  setNewYear,
+  setNewGenre,
+  setNewRating,
+  onSubmit,
+  currentTheme,
+}) => {
+  return (
+    <div
+      style={{
+        marginBottom: "24px",
+        padding: "20px",
+        borderRadius: "12px",
+        backgroundColor: currentTheme.componentBg,
+        border: `1px solid ${currentTheme.border}`,
+      }}
+    >
+      <h2>영화 추가</h2>
+      <form
+        onSubmit={onSubmit}
+        style={{ display: "flex", gap: "12px", flexWrap: "wrap" }}
+      >
+        {/* 영화를 추가하는 기능을 담당하는 부분 */}
+        <input
+          type="text"
+          placeholder="제목"
+          value={newTitle}
+          onChange={(e) => setNewTitle(e.target.value)}
+          style={{
+            padding: "8px",
+            borderRadius: "8px",
+            border: `1px solid ${currentTheme.border}`,
+            backgroundColor: currentTheme.inputBg,
+            color: currentTheme.text,
+          }}
+        />
+        <input
+          type="text"
+          placeholder="감독"
+          value={newDirector}
+          onChange={(e) => setNewDirector(e.target.value)}
+          style={{
+            padding: "8px",
+            borderRadius: "8px",
+            border: `1px solid ${currentTheme.border}`,
+            backgroundColor: currentTheme.inputBg,
+            color: currentTheme.text,
+          }}
+        />
+        <input
+          type="number"
+          placeholder="연도"
+          value={newYear}
+          onChange={(e) => setNewYear(Number(e.target.value))}
+          style={{
+            padding: "8px",
+            borderRadius: "8px",
+            border: `1px solid ${currentTheme.border}`,
+            backgroundColor: currentTheme.inputBg,
+            color: currentTheme.text,
+            width: "80px",
+          }}
+        />
+        <input
+          type="text"
+          placeholder="장르"
+          value={newGenre}
+          onChange={(e) => setNewGenre(e.target.value)}
+          style={{
+            padding: "8px",
+            borderRadius: "8px",
+            border: `1px solid ${currentTheme.border}`,
+            backgroundColor: currentTheme.inputBg,
+            color: currentTheme.text,
+          }}
+        />
+        <input
+          type="number"
+          placeholder="평점"
+          value={newRating}
+          onChange={(e) => {
+            const val = Number(e.target.value);
+            if (val >= 0 && val <= 10) setNewRating(val);
+          }}
+          style={{
+            padding: "8px",
+            borderRadius: "8px",
+            border: `1px solid ${currentTheme.border}`,
+            backgroundColor: currentTheme.inputBg,
+            color: currentTheme.text,
+            width: "100px",
+          }}
+        />
+        <button
+          type="submit"
+          style={{
+            padding: "8px 16px",
+            backgroundColor: currentTheme.buttonBg,
+            color: currentTheme.buttonText,
+            border: "none",
+            borderRadius: "8px",
+            cursor: "pointer",
+          }}
+        >
+          추가
+        </button>
+      </form>
+    </div>
+  );
+};
+
+export default MovieForm;
